[MOD] type() optimization reverted to stay compatible with older version

+++ src/main/java/org/basex/query/value/item/Item.java
@@ -288,14 +288,14 @@ public abstract class Item extends Value {
    * @return type string
    */
   public byte[] xdmInfo() {
-    return new byte[] { typeId() };
+    return new byte[] { (byte) typeId() };
   }
 
   /**
    * Returns a type id.
    * @return type string
    */
-  public byte typeId() {
+  public int typeId() {
     return type.id();
   }
