+++ src/main/java/org/basex/query/value/node/FDoc.java
@@ -112,7 +112,7 @@ public final class FDoc extends FNode {
   }
 
   @Override
-  public byte typeId() {
+  public int typeId() {
     // check if a document has a single element as child
     return children.size() == 1 && children.get(0).type == NodeType.ELM ?
       NodeType.DEL.id() : type.id();
